Move most code out of platform-specific directory.
Much of this may be used in future platforms anyways, but either way it is best to keep the stuff that is actually platform-specific small and isolated. Even if the code only happens to be used on one platform. It also lowers the barrier to using it in a second platform if that makes sense, rather than encouraging writing more code in the platform-specific directory.

Additionally, this makes it much easier to test and work on the code from other platforms as the code compiles.

At this point all that is remaining is the static manager object and the entrypoints that use it by default. If we move this object to the Swift side and pass by pointer we could probably remove all platform-specific code from the Rust side other than the logging setup. But we can do that in the future.