Exit rather than abort.
Primarily this avoids macOS tracking our consecutive crashes and doing something that we don't want based on it. (launchd claims that it is delaying our next start but there seems to be a 5s delay no matter what.) This may also be slighly faster, but it is in the noise and avoids generating an unnnecessary crash report (which may save a bit of compute and possibly speed up the process if the machine is heavily loaded).

Fixes: https://linear.app/soveng/issue/OBS-896/evaluate-exit0-vs-abort-for-restarting-the-ne