Add support for outdated error repsonse.
This response can be used to give good UX when a user with an old client attempts to use a removed API. There are currently no uses of this API in the backend but it is valuable to get into the client now so that as long of a tail of users can have a better error message.

For now this is only shown in a few places due to our adhoc error handling:
1. The login page.
2. Exit list refreshing.

Improving error handling so that this is shown in all error handlers and notifications is tracked by https://linear.app/soveng/issue/OBS-904/improve-error-handling-in-js-client.

This starts that process by adding a dedicated error type for errors that are returned by IPC calls so that these error codes can be identified reliably. It also provides a helper for getting the right translation key with a namespace.

Fixes: https://linear.app/soveng/issue/OBS-901/handle-unsupported-error-in-client