app: persist badge colour when account field is reset to null
When a connection fails and the badge is shown, the account field of the NEStatus / ClientState is reset to null. This is because the NE has to restart due to macOS bugs. This reset casues a crash because the we assumed that a set badge string implies that the account field in the status is not null. To avoid this crash, we make the badge colour a state variable that is computed alongside the badge text. A nil check for account is added to ensure that when the account is reset to null, the badge will remain.