Redact AccountId in logs.
This has a few main points.

1. Use rich types for sensitive data that have safe `Debug` implementations.
2. Print all requests and repsonses on Rust side.
3. Also print the SHA1 hash of the request as macOS frameworks log this.
4. Use the same has for the response for easy correlation.
5. Stop logging request and response bodies outside of the NE.

The main downside to this approach is that if a request fails to get to the NE we won't log the contents. This is likely note a major issue as the fact that requests are failing is likely far more interesting than the specific request body.

It may be nice in the future to generate unique request IDs (or just put some ignored ID into the request body before hashing) so that the requests and responses of identical requests can be identified separately, but that can easily be done in the future.

Fixes: https://linear.app/soveng/issue/OBS-895/dont-show-account-id-in-ne-command-success-log