Fix type of pinned exits.
This causes the Swift to fail to parse the status. This likely wasn't noticed because a failed status update is ignored (as failing due to NE restart is common) and the web UI correclty handles it. So this would only be noticed if you had a non-migrated pinned exit on startup in which case the app would fail to initialize.

Fixes: https://linear.app/soveng/issue/OBS-1203/angel-h-problem-initializing-tunnel-provider-v1112