Ping Obscura API in the debug bundle.
We have had some people who seemed to be connected to the internet but couldn't connect to our API. This will provide a bit of insight into this as well as clearly indicate which backend commit was running (when the bundle was generated).

I have disabled cert checking for the request as it prints the validation status either way and I would like to see the result even if the certificate isn't trusted for some reason.

Fixes: https://linear.app/soveng/issue/OBS-1449/add-basic-ping-endpoint-to-api-server-for-debug-bundle