Add option to use `includeAllNetworks`.
In OS-independant parts of the code I called this `strictLeakPrevention`.

My main concern here is that we are just tracking the macOS property. I'm worried that if saving fails we may show the wrong value. Probably the best option would be to have our own vairable that we track before attempting changes and clear only when we are sure we know the current state again. But that can be added before shipping the end-user UI, for now at least the core functionality works.

Bug: https://linear.app/soveng/issue/OBS-234/consider-setting-the-includeallnetworks-network-extension-flag