Cache exit list.
This implements the general pattern that we discussed for data in the NYC meetup. This pattern as it is expected to be reused by at least relays, but probably also other data.

1. Use a subscription like we do for status.
2. Provide a method that requests a refresh up to a given staleness threshold.

On top of that this implements some improvements:

1. Persist exit list between executions.
2. Remove exit list context to avoid re-evaluating the entire app whenever the exit list changes.
3. Refresh the exit list every 60s when there is a subscriber. This should be fairly cheap due to the caching. If there is no subscribers (a component that is displaying the data) the exit list will no longer be refreshed.

The main motivation for this change is that it makes the exit list persistently available to the network extension so that we can move exit selection there.

Support for migrating the old pinned exit list was also removed as the could would need to be updated and this migration has been in place for a long time and the data that would be lost is not very valuable.

Bug: https://linear.app/soveng/issue/OBS-1633/add-exits-and-relays-caching-to-client