Don't rotate wireguard key on handshake errors.
We frequenlty see config propagation errors and rotating the key doesn't help and it puts more load on Mullvad, and specifically the system that is having trouble.

It seems that any dead key error is likely an issue on our side and it would be better to fix it than paper over it.

Fixes: https://linear.app/soveng/issue/OBS-1666/stop-auto-rotating-wireguard-keys