Cache exit list in client.
Since the exit list is notably part of the UI we want it to be available instantly. Even fetching from the NE can take enough time that there is a noticable delay where the exits are missing. This updates to use a shared cache on the fontend so that any new component can access a ready-to-go exit list.

In addition since the cache was being made cross-component it made sense to only actually fetch the list once. This saves creating lots of objects in each component and lowers memory usage as there is only one copy in memory in most cases.