Fix missing status item on macOS 15
On macOS 15, apple changed the behavior of login item apps such that the Window of login item apps will not appear. They also fixed the crash if the status item was created before onAppear (as in on App init). Decided to move the initialization to AppDelegate which does away with conditional compilation.