Use aws-lc-rs
This is now the default in rustls but some work was needed to make it work on iOS.

1. Install cmake.
2. Enable the `bindgen` feature.

For installing cmake we will continue with the tradition of just using everything on the system. Eventually it would be nice to do the whole build in Nix but it isn't a priority right now. Since we are installing cmake I updated the instructions to install both rustup and cmake using Nix, this removes one step. Installing rustup via homebrew or its own installation script should continue working.

Fixes: https://linear.app/soveng/issue/OBS-1885/fix-nix-aws-lc-rs-builds