Ruduce the amount and cost of keepalives.
Previously we would always send keepalives on a 1s interval. This reduces that to the following scenarios.

1. If we transmit but don't recieve a response within 1s.
2. If we don't send anything for 60s.
3. If we send data, and haven't transmitted a keepalive in 5min.

This means that as long as the tunnel is being used reguarly the only extra data we send is a single keepalive every ~5min which is sent at the same time as user traffic (so the radio should already be awake).

For completely idle connections we will send and recieve one packet every minute. This could likely be reduced further but 1min seemed like a good starting point. We do want to eventually notice breaks even if nothing is being transmitted in case the other end is trying to send something. Additionally we want to keep any NAT leases alive.

The remaining case should be rarely seen in practice as almost all internet protocols have some sort of acknoledgement to ensure that data is transmitted. However if the user does send data that isn't acknoledged we will send out our own acknoledgement/keepalive to ensure that the remote end isn't dead.

Fixes: https://linear.app/soveng/issue/OBS-2096/reduce-keepalive-pings-on-client