Limit time waiting for additional relay connections.
Previously the code would wait for the first 5 responses. This was usually fine but in cases where the user has a bad connection to many relays or some relays are offline (including cases where only one port is working for a user) this could result in very long connection times.

Now, once we have one candidate we limit the total wait to 5s or 3x the time it took to discover the current best candidate (whichever is shorter). This is because a candidate that takes 3x longer to connect is very unlikely to be faster and it is better to just get the user connected than wait for longer than 5 seconds.

For now there is no minimal connection time. For me testing I was frequently able to get 90ms deadlines which resulted in noticably faster connection times. We may consider a minimal time in the future (50ms?) but honestly if the user can connect to a relay in under 17ms with 4-5 round trips their ping is a few ms anyways and I don't think looking for something faster adds much value. So for now I put no minimal wait time.

The limit of 5 unique relays is mostly obsolete now. However at the cost of a few allocations and comparisons it doesn't hurt to keep it in. The main risk is that it affects staging significantly differently since there are only 3 relays so we will never hit the limit there, effectively giving relay selection different behaviour in different environments.

Fixes: https://linear.app/soveng/issue/OBS-2102/relay-selection-shouldnt-wait-for-slow-relays