Clean up iOS deployment target.
The previous commit was the simplest change that worked so it was used to build the release. We don't want to ammend it to preserve the exact build commit in the history. This commit simplifies the change to the minimum necessary.

I validated this change by building and uploading to TestFlight and checking the minimum iOS version listed (there must be an easier way...).